requirements.txt
mcp[cli]
requests
google-cloud-aiplatform>=1.38.0
vertexai>=0.0.1
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
